refactor: move the currently component to a separate folder.
